Issue 124: Purchase Order Workflow: Add email task to end of purchase order workflow.
Append to the workflow building and sending e-mail tasks.

Utilize dynamic and configurable file path.
The `path` json property no longer expects the path to the file and now expects the path to the directory containing the file.

The e-mail task is setting the text as HTML markup, but this does not work well with plain text (specifically the newlines).
Such an e-mail is not easily readable.
Change the behavior to send both plain text and HTML markup in a backwards compatible manner.
This introduces a new field called `markup`/`mailMarkup` that exists optionally in addition to the field `text`/`mailText`.