Issue 103: Eresources/Cataloging Workflow: Create brief instance records for new Coral items.
This approach performs the entire process in one setting rather than breaking them apart across days (which would suggest separate workflows and maintain files on the disk).

Inclusive Gateways are used to determine when an Instance or Holdings is to be or is not to be created.
The Instance and Holdings should only be created if they do not already exist.

*Special Notes*

There are two distinct ways in which regex is being parsed in the Javascript files:

1)
```
.replace(/\[title\]/i, safe(item.title))
```

2)
```
var match = new RegExp('\\[instanceId\\]', 'i');
...
.replace(match, instanceId)
```

The reason for this is because having `instanceId` outside of a string while there is an autogenerated variable called `instanceId` results in the variable substituting into a string rather than the regex being the string "instanceId".
To prevent this behavior, the `new RegExp(...` logic is used to prevent the string "instanceId" from being treated as a variable.