Issue 307: POWF - to make the expenseClass optional.
The workflow is failing due to not properly handling existence cases, such as NULL checks.
Provide a helper function `extractResponseArray()` to perform the needed checks.

This change does not determine what to do when required values are not defined.
An error at a different part of the process may occur but such an error will likely be due to bad data.

The decision here is to use `undefined` rather than `null` for the cases where the existence checks fail.
If this is a problem, then this can be changed to `null` instead of `undefined`.