Issue 308: Update cron job, the Camunda cron is different from standard cron.
The first number is not minute, it is second.
This caused the cron job to run on the 7th minute of every hour rather at 7 am on every day.

Note: The last value apparently must be a `?` and not `*` or mod-camunda will fail on:
```
java.text.ParseException: Support for specifying both a day-of-week AND a day-of-month parameter is not implemented.
```