Issue 308: Change LDP select query to use more accurate filtering.
The previous SQL utilized `any()`.
This is apparently a mistake.
The `any()` will select anything that matches.
In the case where there are two things in the array, one that matches and one that doesn't, then the `any()` will match.
This is not the correct behavior.

The new approach is to use the `jsonb` contains operator `<@`.
The LDP uses `json` and not `jsonb` like Folio.
The Postgresql decided not to provide support for the contains operator for `json`.
The solution here is to cast the `circulationNotes` array into a `jsonb` structure.
This likely has a performance impact but it is after the other filters are selected.
This should keep the cost down.
The `circulationNotes` will likely normally be a small array.

Note: The syntax `<@` was attempted but the Camunda Freemarker syntax tries to interpret this.
Change the `<@` into `\l@` then causes the JSON to have an escape problem for the `\`.
Change the `\l@` into `\\l@` and somehow the should be resulting `\l` does not get processed correctly by freemarker and it errors on the slash.
The solution is to instead swap the left and right hand sides using instead `@>`.

This is the SQL query:
```sql
(
  select id from inventory_items
    where effective_location_id = '07a4e97e-9941-4f60-ad25-577bb6672c08' and status__name = 'Checked out' and json_array_length(data->'circulationNotes') = 0
) union (
  select s.id from (
    select id, to_jsonb(data->'circulationNotes') as notes
    from inventory_items where effective_location_id = '07a4e97e-9941-4f60-ad25-577bb6672c08' and status__name = 'Checked out'
  ) s
    where not s.notes @> '[{"note":"Place on New Bookshelf"}]'::jsonb
)
```