Issue 316: Add Create Notes Workflow.
This fetches first from LDP to get a fine-tuned small list.
Then does a follow up fetch to make sure there are no duplicate notes added.

Both the LDP variables and the OKAPI variables are selected and request time to avoid saving username, password, and URLs in the built JSON file.
The security of the credentials is expected to be handled by Camunda as per discussion with the Product Owner.

The current behavior of handling `staffOnly` is to perform a complete exact object match of the complete message object (**UUID**, `note`, and `staffOnly`).
If a message with an identical **UUID** and an identical `note` message have a different `staffOnly` property, then a new message is created.
The alternative behavior that should be considered before approving these changes is whether or not the identical **UUID** and identical `note` message should update if the `staffOnly` is different.

Multiple **TODO** comments are added concerning `StringUtils.escapeSql()`.
The `StringUtils.escapeSql()` is currently failing for unknown reasons. For the sake of time, this is being designate as a **TODO**.
The `try..catch..` block is left in place should the `StringUtils.escapeSql()` become implemented.