Fix problems in the README file.
Make the README structure slightly more consistent.
This only addresses some, but not all, of the inconsistency problems.

Improve the grammar and language being used.

Use `shell` more consistently to better improve presentation on render.

The examples showing `X-Okapi-Tenant:` are inconsistent.

The documentation describing the files and paths in some cases are incorrect.

Make the variable table more consistently used.
I had to glance at the requirements and so there may be more things to fix regarding this.
In some cases I discovered that the examples are entirely wrong and reference incorrect or non-existent variables.

Get rid of TAMU-specific information.
The examples should just be examples and should not be tied to some institution or system.

Use alphabetical ordering in the table to make searching the variable names easier.