UIWRKFLOW-12: Ensure example workflows are in a consistent and working state.
These Workflows will be zipped into an FWZ and provided as an example to use in the Scratch environment.

Some example Workflows have `Workflow` in the name and some do not.
Make them consistent and have `Workflow` in the name.

Remove unused `buildResponse.json` that is causing problems.

The `Compress File` should also have `Example` in the name.
Also add the compress type `(ZIP)` in the name.

The URL request response should generally be `text/html`.